improve argument handling and add style-attribute replacement

+++ pngs_from_svg.py
@@ -1,18 +1,65 @@
 #!/usr/bin/python3
 
-import sys
 import os
 import shlex
 import subprocess
 import tempfile
 import shutil
+import argparse
+import re
+
+import xml.etree.ElementTree as ElementTree
 
 DIRS = ["drawable-mdpi", "drawable-hdpi", "drawable-xhdpi", "drawable-xxhdpi"]
 
+def remove_color(path):
+    style = path.get('style')
 
-def create_images(svg, dir_pngs, namebase, suffix, isize, color, opacity):
+    if style is not None:
+        style = re.sub("fill:\s*#\d{3,6};?", "", style)
+        path.set("style", style)
+
+
+def modify_svg(svg, tmp, color, opacity):
+    ns = "http://www.w3.org/2000/svg"
+    ElementTree.register_namespace("", ns)
+
+    tree = ElementTree.parse(svg)
+    root = tree.getroot()
+
+    if color:
+        for path in tree.iter("{" + ns + "}path"):
+            if path.get('fill') != "none":
+                remove_color(path)
+                path.set("fill", color)
+
+    if opacity:
+        opacityGroup = ElementTree.Element("g")
+        opacityGroup.set("id", "png_from_svg")
+        opacityGroup.set("opacity", str(opacity))
+        root.append(opacityGroup)
+
+        ignoredElementTags = ['{http://www.w3.org/2000/svg}metadata',
+                              '{http://www.w3.org/2000/svg}defs',
+                              '{http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd}namedview']
+
+        for path in root.findall('*'):
+            if (path != opacityGroup) and (path.tag not in ignoredElementTags):
+                opacityGroup.append(path)
+                root.remove(path)
+
+    tree.write(tmp)
+
+
+def create_images(svg, dir_pngs, namebase, suffix, isize, color, opacity, nopngcrush=False):
     # dimensions for mdpi/hdpi/xhdpi/xxhdpi
     sizes = [isize, isize * 1.5, isize * 2, isize * 3]
+
+    tmp_svg = tempfile.mkstemp(".svg")
+    os.close(tmp_svg[0])
+
+    modify_svg(svg, tmp_svg[1], color, opacity)
+
     for s in range(len(sizes)):
         size = sizes[s]
         output_dir = os.path.join(dir_pngs, DIRS[s])
@@ -23,54 +70,27 @@ def create_images(svg, dir_pngs, namebase, suffix, isize, color, opacity):
         output = os.path.join(output_dir, namebase + suffix + ".png")
         print (svg + ' -> ' + output)
 
-        tmp_svg = tempfile.mkstemp(".svg")
-        os.close(tmp_svg[0])
+        create_png(tmp_svg[1], output, size, nopngcrush)
 
-        f = open(svg, 'r')
-        data = f.read()
-        f.close()
-
-        data = data.replace(
-            "style=\"fill:#000000;opacity:1\"",
-            "style=\"fill:" + color + ";opacity:" + str(opacity) + "\"")
-        data = data.replace( \
-            "style=\"opacity:1;fill:#000000\"",
-            "style=\"opacity:" + str(opacity) + ";fill:" + color + "\"")
-        data = data.replace(
-            "fill=\"#000000\"",
-            "fill=\"" + color + "\"")
-        data = data.replace(
-            "style=\"fill:#000000\"",
-            "style=\"fill:" + color + "\"")
-        data = data.replace(
-            "fill-opacity=\"1.0\"",
-            "fill-opacity=\"" + str(opacity) + "\"")
-
-        f = open(tmp_svg[1], 'w')
-        f.write(data)
-        f.close()
-
-        create_png(tmp_svg[1], output, size)
-
-        os.remove(tmp_svg[1])
+    os.remove(tmp_svg[1])
 
 
-def create_png(svg, output, size):
+def create_png(svg, output, size, nopngcrush):
     cmd = "inkscape -C -e " + output + " -h " + str(size) + " " + svg
     args = shlex.split(cmd, False, running_on_posix)
     subprocess.call(args)
 
-    tmp_png = tempfile.mkstemp(".png")
-    os.close(tmp_png[0])
-    shutil.copyfile(output, tmp_png[1])
-    os.remove(output)
+    if nopngcrush == False:
+        tmp_png = tempfile.mkstemp(".png")
+        os.close(tmp_png[0])
+        shutil.copyfile(output, tmp_png[1])
+        os.remove(output)
 
-    cmd = "pngcrush " + tmp_png[1] + " " + output
-    args = shlex.split(cmd, False, running_on_posix)
-    print (args)
-    subprocess.call(args)
+        cmd = "pngcrush " + tmp_png[1] + " " + output
+        args = shlex.split(cmd, False, running_on_posix)
+        subprocess.call(args)
 
-    os.remove(tmp_png[1])
+        os.remove(tmp_png[1])
 
 
 def is_posix():
@@ -80,30 +100,23 @@ def is_posix():
     except ImportError:
         return False
 
-
 running_on_posix = is_posix()
 
 if __name__ == "__main__":
 
-    nargs = len(sys.argv)
-    if nargs != 6 and nargs != 7:
-        print ("usage: " + sys.argv[0] +
-               " <filename (input)> <res directory (output)> <size> <color> <opacity> [<suffix>]")
-        exit(1)
+    parser = argparse.ArgumentParser()
+    parser.add_argument("input_filename",      help="Path to the svg file")
+    parser.add_argument("output_directory",    help="Output directory")
+    parser.add_argument("size",                help="Size in pixel for the mdpi variant", type=int)
+    parser.add_argument("-c", "--color",       help="Color, e.g. '#104E8B'")
+    parser.add_argument("-o", "--opacity",     help="Opacity, e.g. '0.54'",               type=float)
+    parser.add_argument("-s", "--suffix",      help="Suffix for the output filenames",    default="")
+    parser.add_argument("-nc", "--nopngcrush", help="Don't run pngcrush", action="store_true")
+    args = parser.parse_args()
 
-    svg_path = sys.argv[1]
-    dir_pngs = sys.argv[2]
-    isize = int(sys.argv[3])
-    color = sys.argv[4]
-    opacity = float(sys.argv[5])
-
-    suffix = ""
-    if nargs == 7:
-        suffix = sys.argv[6]
-
-    svg_name = os.path.basename(svg_path)
+    svg_name = os.path.basename(args.input_filename)
     name = svg_name
     if svg_name.lower().endswith(".svg"):
         name = svg_name[:-4]
 
-    create_images(svg_path, dir_pngs, name, suffix, isize, color, opacity)
+    create_images(args.input_filename, args.output_directory, name, args.suffix, args.size, args.color, args.opacity, nopngcrush=args.nopngcrush)
